.claude/skills/backlog/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.18.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
